Moved the view selector to the common components.
