+++ src/javascript/query-explorer/components/query-explorer.js
@@ -18,14 +18,13 @@ import accountSummaries from 'javascript-api-utils/lib/account-summaries';
 
 import QueryReport from './query-report';
 import Select2MultiSuggest from './select2-multi-suggest';
-import ViewSelector from './view-selector';
 
 import {gaAll} from '../../analytics';
 import AlertDispatcher from '../../components/alert-dispatcher';
 import Datepicker from '../../components/datepicker';
 import HelpIconLink from '../../components/help-icon-link';
 import SearchSuggest from '../../components/search-suggest';
-
+import ViewSelector from '../../components/view-selector';
 
 /**
  * The parameters that are safe to track the values entered by users.
