+++ src/javascript/request-composer/components/request-composer.js
@@ -20,7 +20,6 @@ import accountSummaries from 'javascript-api-utils/lib/account-summaries';
 import {composeRequest} from '../request';
 import RequestViewer from './request-viewer';
 import Select2MultiSuggest from './select2-multi-suggest';
-import ViewSelector from './view-selector';
 import ResponseViewer from './response-viewer';
 
 import {gaAll} from '../../analytics';
@@ -28,7 +27,7 @@ import AlertDispatcher from '../../components/alert-dispatcher';
 import Datepicker from '../../components/datepicker';
 import HelpIconLink from '../../components/help-icon-link';
 import SearchSuggest from '../../components/search-suggest';
-
+import ViewSelector from '../../components/view-selector';
 
 /**
  * The parameters that are safe to track the values entered by users.
